Refactor Icons component - Move static data to iconsData module - Move renderIcon to separate file - Refactor component
