+++ src/data/iconsData.js
@@ -1,0 +1,34 @@
+const iconTitles = {
+  android: 'Android',
+  apple: 'iOS',
+  bootstrap: 'Bootstrap',
+  css3: 'CSS3',
+  coffee: 'Developer - An organism that turns coffee into code.',
+  django: 'Django',
+  expo: 'Expo',
+  gatsby: 'Gatsby',
+  graphql: 'GraphQL',
+  gulp: 'Gulp',
+  highcharts: 'Highcharts',
+  html5: 'HTML5',
+  jquery: 'jQuery',
+  js: 'JavaScript',
+  node: 'Node',
+  php: 'PHP',
+  'post-css': 'PostCSS',
+  postgresql: 'PostgreSQL',
+  python: 'Python',
+  react: 'React',
+  'react-native': 'React Native',
+  remix: 'Remix',
+  sass: 'Sass',
+  'styled-components': 'Styled Components',
+  tailwind: 'Tailwind CSS',
+  ts: 'TypeScript',
+  vite: 'Vite',
+  vuejs: 'Vue',
+  webpack: 'Webpack',
+  wordpress: 'WordPress',
+};
+
+export default { iconTitles };
